+++ LinearRegression.h
@@ -6,12 +6,12 @@ public:
     Regression(void);
     Regression(const Regression &copy);
     void Reset(void);
-    void Append(float sample);
-    float GetSlope(void);
-    float GetOffset(void);
+    void Append(double sample);
+    double GetSlope(void);
+    double GetOffset(void);
 private:
-    float slope, offset;
-    float b0, b1;
+    double slope, offset;
+    double b0, b1;
     int n;
 };
