+++ SOS.cpp
@@ -1,10 +1,10 @@
 #include "SOS.h"
 
-float kernel(float in, const struct SOSystem &coef, float * mem){
+double kernel(double in, const struct SOSystem &coef, double * mem){
     in *= coef.a0;
     in -= mem[0] * coef.a1;
     in -= mem[1] * coef.a2;
-    float out = in * coef.b0;
+    double out = in * coef.b0;
     out += mem[0] * coef.b1;
     out += mem[1] * coef.b2;
     mem[1] = mem[0];
@@ -12,22 +12,22 @@ float kernel(float in, const struct SOSystem &coef, float * mem){
     return out;
 }
 
-float SOSfilter::Process(float sample){
-    float y = 1;
+double SOSfilter::Process(double sample){
+    double y = 1;
     for(int i = 0; i < len; ++i){
         y *= kernel(sample, h[i], x+i+i);
     }
     return y;
 }
 
-float SOSfilter::Stability(void){
+double SOSfilter::Stability(void){
     //FIXME: calculate stability
     return 1;
 }
 
 /** FIXME: I don't think this works properly */
-void SOSfilter::Reset(float val=0.){
-    float a = Stability();
+void SOSfilter::Reset(double val=0.){
+    double a = Stability();
     // Calculate the steady-state condition of a one-element feedback
     // steady = input + steady*(a) ===> steady = input / (1-a)
     val = val / (1-a);
@@ -42,7 +42,7 @@ SOSfilter::SOSfilter(const struct SOSystem * coef, uint8_t size){
     len = size / (sizeof(struct SOSystem));
     h = coef;
     int len2 = len * 2;
-    x = new float[len2];
+    x = new double[len2];
     for(int i = 0; i < len2; ++i){
         x[i]=0;
     }
@@ -52,7 +52,7 @@ SOSfilter::SOSfilter(const SOSfilter &copy){
     len = copy.len;
     h = copy.h;
     int len2 = len * 2;
-    x = new float[len2];
+    x = new double[len2];
     for(int i = 0; i < len2; ++i){
         x[i]=0;
     }
