+++ SOS.h
@@ -4,12 +4,12 @@
 #include <inttypes.h>
 
 struct SOSystem{
-    float b0;
-    float b1;
-    float b2;
-    float a0;
-    float a1;
-    float a2;
+    double b0;
+    double b1;
+    double b2;
+    double a0;
+    double a1;
+    double a2;
 };
 
 /* cascaded Second Order System */
@@ -17,15 +17,15 @@ class SOSfilter{
 public:
     SOSfilter(const struct SOSystem * coeffecients, uint8_t size);
     SOSfilter(const SOSfilter &copy);
-    void Reset(float val);
-    float Process(float sample);
-    float Stability(void);
+    void Reset(double val);
+    double Process(double sample);
+    double Stability(void);
     ~SOSfilter(void);
     friend class ZPK;
 private:
     const struct SOSystem * h; // array of second-order coeffecients
     uint8_t len;
-    float * x;
+    double * x;
 };
 
 #endif
